+++ ICD.Connect.Routing.CrestronPro/Transmitters/DmTx4kX02CBase/AbstractDmTx4kX02CBaseAdapter.cs
@@ -13,7 +13,6 @@ using ICD.Connect.API.Nodes;
 #endif
 using ICD.Common.Properties;
 using ICD.Common.Utils;
-using ICD.Common.Utils.Extensions;
 using ICD.Connect.Routing.Connections;
 
 namespace ICD.Connect.Routing.CrestronPro.Transmitters.DmTx4kX02CBase
@@ -32,8 +31,6 @@ namespace ICD.Connect.Routing.CrestronPro.Transmitters.DmTx4kX02CBase
 		protected const int HDMI_INPUT_2 = 2;
 		protected const int DM_OUTPUT = 1;
 		protected const int HDMI_OUTPUT = 2;
-
-		private bool m_ActiveTransmissionState;
 
 		#region Properties
 
@@ -54,32 +51,10 @@ namespace ICD.Connect.Routing.CrestronPro.Transmitters.DmTx4kX02CBase
 			}
 		}
 
-		/// <summary>
-		/// Returns true when the device is actively transmitting video.
-		/// </summary>
-		[PublicAPI]
-		public bool ActiveTransmissionState
-		{
-			get { return m_ActiveTransmissionState; }
-			protected set
-			{
-				if (value == m_ActiveTransmissionState)
-					return;
-
-				m_ActiveTransmissionState = value;
-
-				RaiseActiveTransmissionStateChanged(HDMI_OUTPUT,
-													  eConnectionType.Audio | eConnectionType.Video,
-													  m_ActiveTransmissionState);
-			}
-		}
-
 		#endregion
 
 		#region Methods
 
-#if !NETSTANDARD
-#endif
 		/// <summary>
 		/// Returns true if a signal is detected at the given input.
 		/// </summary>
@@ -199,32 +174,6 @@ namespace ICD.Connect.Routing.CrestronPro.Transmitters.DmTx4kX02CBase
 			yield return new ConnectorInfo(HDMI_OUTPUT, (eConnectionType.Audio | eConnectionType.Video));
 		}
 
-		public override bool GetActiveTransmissionState(int output, eConnectionType type)
-		{
-			if (EnumUtils.HasMultipleFlags(type))
-			{
-				return EnumUtils.GetFlagsExceptNone(type)
-								.Select(f => GetActiveTransmissionState(output, f))
-								.Unanimous(false);
-			}
-
-			if (!ContainsOutput(output))
-			{
-				string message = string.Format("{0} has no {1} output at address {2}", this, type, output);
-				throw new ArgumentOutOfRangeException("output", message);
-			}
-
-			switch (type)
-			{
-				case eConnectionType.Audio:
-				case eConnectionType.Video:
-					return ActiveTransmissionState;
-
-				default:
-					throw new ArgumentOutOfRangeException("type");
-			}
-		}
-
 		/// <summary>
 		/// Performs the given route operation.
 		/// </summary>
@@ -302,7 +251,14 @@ namespace ICD.Connect.Routing.CrestronPro.Transmitters.DmTx4kX02CBase
 			return false;
 		}
 
-		protected virtual bool GetActiveTransmissionState()
+		/// <summary>
+		/// Gets the source detect state from the Tx
+		/// This is just a simple "is a laptop detected"
+		/// Used for ActiveTransmission in AutoRouting mode
+		/// Override to support additional inputs on a Tx
+		/// </summary>
+		/// <returns></returns>
+		protected override bool GetSourceDetectionState()
 		{
 			return HdmiDetected;
 		}
@@ -385,7 +341,7 @@ namespace ICD.Connect.Routing.CrestronPro.Transmitters.DmTx4kX02CBase
 		{
 			if (args.EventId == EndpointInputStreamEventIds.SyncDetectedFeedbackEventId)
 			{
-				ActiveTransmissionState = GetActiveTransmissionState();
+				UpdateSourceDetectionState();
 				SwitcherCache.SetSourceDetectedState(HDMI_INPUT_1, eConnectionType.Video,
 													 Transmitter.HdmiInputs[HDMI_INPUT_1].SyncDetectedFeedback.BoolValue);
 				SwitcherCache.SetSourceDetectedState(HDMI_INPUT_1, eConnectionType.Audio,